+++ src/main/java/com/example/newsfeedproject/common/annotation/LoginUserResolver.java
@@ -1,4 +1,4 @@
-package com.example.newsfeedproject.common.annoation;
+package com.example.newsfeedproject.common.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
